+++ spring-petclinic-api-gateway/src/main/resources/application.yml
@@ -25,8 +25,6 @@ spring:
           filters:
             - StripPrefix=2
 ---
-server:
-  port: 8080
 spring:
   config:
     activate:
@@ -34,11 +32,28 @@ spring:
   cloud:
     gateway:
       routes:
+        - id: vets-service
+          uri: lb://vets-service
+          predicates:
+            - Path=/api/vet/**
+          filters:
+            - StripPrefix=2
+        - id: visits-service
+          uri: lb://visits-service
+          predicates:
+            - Path=/api/visit/**
+          filters:
+            - StripPrefix=2
+        - id: customers-service
+          uri: lb://customers-service
+          predicates:
+            - Path=/api/customer/**
+          filters:
+            - StripPrefix=2
         - id: frontend
           uri: http://localhost:3000
           predicates:
             - Path=/**
-
 ---
 server:
   port: 8080
