+++ spring-petclinic-customers-service/src/main/resources/application.yml
@@ -4,13 +4,6 @@ spring:
   config:
     import: optional:configserver:${CONFIG_SERVER_URL:http://localhost:8888/}
 
----
-server:
-  port: 8081
-spring:
-  config:
-    activate:
-      on-profile: development
 
 ---
 spring:
