+++ spring-petclinic-frontend/.env.production
@@ -1,4 +1,1 @@
 NODE_ENV=production
-CUSTOMERS_SERVICE=http://customers-service
-VETS_SERVICE=http://vets-service
-VISITS_SERVICE=http://visits-service
