+++ spring-petclinic-frontend/package.json
@@ -21,7 +21,6 @@
   "devDependencies": {
     "copy-webpack-plugin": "^11.0.0",
     "css-loader": "^6.8.1",
-    "dotenv-webpack": "^8.0.1",
     "fs-extra": "^11.1.1",
     "html-loader": "^4.2.0",
     "html-webpack-plugin": "^5.5.3",
@@ -34,5 +33,7 @@
     "webpack-dev-server": "^4.15.1",
     "webpack-merge": "^5.10.0",
     "webpack-shell-plugin-next": "^2.3.1"
-  }
+  },
+  "main": "webpack.config.js",
+  "keywords": []
 }
