+++ spring-petclinic-frontend/src/scripts/fragments/footer.html
