+++ spring-petclinic-frontend/src/scripts/fragments/welcome.html
