+++ spring-petclinic-frontend/src/scripts/vet-list/vet-list.js
