+++ spring-petclinic-frontend/src/scripts/visits/visits.template.html
